ntp time fetching - not get 48 bytes of response

+++ RxModuleTI/time_lib.h
@@ -17,7 +17,7 @@ extern "C" {
 
 #include "stdint.h"
 
-typedef uint64_t timeData_t;
+typedef uint32_t timeData_t;
 
 typedef enum {timeNotSet, timeNeedsSync, timeSet
 }  timeStatus_t ;
